refactor: refactored dark mode toggle to separate component from footer component. removed and relocated toggle to navbar for better user intuitiveness.
